Enhancements to Windows service install/uninstall.
If --reg-winsvc-auto-start is specified when installing the service
(--reg-winsvc i), set the service type to AUTO_START rather than
DEMAND_START, and also start the service immediately.

When uninstalling (--reg-winsvc u), attempt to stop the service if it
is running.  This is done unconditionally.

Also allow --reg-winsvc-fluentdopt to be specified along with --reg-winsvc;
previously the options would be ignored.  The flag can still be specified
alone to change the options (although a service restart is necessary).