Add nanoseconds to stdout plugin
Now that Fluentd supports nanosecond-resolution timestamps as of v0.14,
it follows that the stdout output plugin should print the full
timestamp.

This is a very simple patch, instead of using the default `to_s`
formatting, we call `strftime` with a custom formatting.

Right now it always prints the nanoseconds, even on second-resolution
timestamps.  This means second-resolution timestamps will display like
`2013-06-27 12:34:56.000000000 -0700`, which may be undesirable.  If so,
we can adjust the implement to print a condensed formatting for
second-resolution timestamps.