Mask all secret parameters in worker section, fix #1553
Secrete parameters are masked when configuring the plugin instances, but fluentd worker processes don't create and configure the instances of plugins under worker sections for other workers. And config dump is performed on a worker process (worker0). So the secret parameters under worker section are not masked except for worker0.
This patch makes supervisor process create and configure instances of plugins for all workers before spawning worker processes and config_dump performed on supervisor process. This method is the same with dry_run.