test: Fix random test failures in test/plugin/test_out_forward.rb
I notice that the following test cases (which are supposed to test the
handling of broken ack responses) are randomly failing on Travis CI:

 1. 'a destination node not supporting responses by disconnection'
 2. 'a destination node not supporting responses by just ignoring'

The root problem is that, since the destination node gets labelled as
"unavailable" after an ack failure, there really shouldn't be any flush
attempt after the testing steps (or it will raise a NoNodesAvailable
exception).

In most environments, this possibility does not pose much problems
because the test execution is sufficiently fast (so there is very
little space for an additional flush attempt to occur).

But since the CI server is known to become very slow from time to time,
we need to take special care for it.