test_out_file.rb: Fix test failures due to Zlib::DataError.
AppVeyor CI tests are constantly failing due to Zlib-related exceptions
like below:

    Error: test: append(FileOutputTest): Zlib::DataError: invalid
    distance too far back

This is a very weird error, because the file seems to contain a proper
gzip-compressed data stream. In fact, by inserting the following code
before the failing line, I can confirm that the file itself is NOT
broken:

    bytes = File.open(path).read   // Read the file being tested
    puts(Zlib::GzipReader.new(StringIO.new(bytes)).read)

The strong suspicion is that Ruby's Zlib wrapper (or its build on
AppVeyor) has a serious bug that corrupts file pointers passed to
GzipReader, thus it ends up reading garbage data.

This patch is my attempt to mitigate the issue by modifying the test
functions to pass bytes (rather than file objects) to GzipReader, and
it should protect zlib from read-time data corruptions.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>